Tighten projected curve search types

+++ include/main/dll/objfsa.h
@@ -80,6 +80,6 @@ int FUN_800df2a4(double param_1,double param_2,double param_3,int param_4,int pa
 uint FUN_800df46c(undefined8 param_1,double param_2,double param_3);
 void curves_lengthFn24(undefined4 param_1,undefined4 param_2,float *param_3,float *param_4);
 void curves_getPos(double phase,int curve,float *outX,float *outY,float *outZ);
-int RomCurve_findProjectedCurveFromStart(double x,double y,double z,int curve,float *outPhase);
+int RomCurve_findProjectedCurveFromStart(f32 x,f32 y,f32 z,int curve,float *outPhase);
 
 #endif /* MAIN_DLL_OBJFSA_H_ */
